feat: 채택된 답변 수정·삭제 차단

+++ questions/views.py
@@ -192,6 +192,13 @@ class AnswerViewSet(
     def update(self, request, *args, **kwargs):
         partial = kwargs.pop("partial", False)
         instance = self.get_object()
+
+        if instance.is_accepted:
+            return error_response(
+                "채택된 답변은 수정할 수 없습니다.",
+                status_code=status.HTTP_400_BAD_REQUEST,
+            )
+
         serializer = self.get_serializer(
             instance, data=request.data, partial=partial, context={"request": request}
         )
@@ -204,6 +211,13 @@ class AnswerViewSet(
 
     def destroy(self, request, *args, **kwargs):
         instance = self.get_object()
+
+        if instance.is_accepted:
+            return error_response(
+                "채택된 답변은 삭제할 수 없습니다.",
+                status_code=status.HTTP_400_BAD_REQUEST,
+            )
+
         instance.delete()
         return success_response(None, "답변이 삭제되었습니다.")
